feat: add removeLru proc that returns removed LRU item
Add exception handling to entire cache.

`del` now returns the value that was deleted, using `Option[T]`. If nothing was deleted, `T.none` is returned.

`removeLru` removes the least recently used item and returns it, or throws an `EmptyLruCacheError` if the cache is empty.

Add tests for `removeLru`.
Modify `del` test for option.

Release 1.2.0